MS-182 refactor(code): code refactoring
