+++ src/store/state.ts
@@ -1,5 +1,5 @@
 import { getItemFromLocalStorage } from "../utils/utils";
-import { CatalogQueryArgs, State } from "../types/types";
+import { State } from "../types/types";
 
 export const state: State = {
   name: getItemFromLocalStorage<string>("user"),
@@ -8,14 +8,6 @@ export const state: State = {
   apiRoot: null,
   cartId: getItemFromLocalStorage<string>("cart-id"),
   anonymousId: getItemFromLocalStorage<string>("anonymousId"),
-};
-
-export const catalogQueryArgs: CatalogQueryArgs = {
-  searchText: null,
-  pageNumber: 1,
-  category: null,
-  filter: null,
-  sort: null,
 };
 
 export default state;
